+++ examples/ending_locations/del_attr.py
@@ -3,3 +3,4 @@ class Foo():
         self.attr = 1
         del self.attr
         del self . attr
+        del self.attr; del self.attr